# Adapt Raytrace_Benchmark example to use BVH acceleration structure