… examples/Raytracer_Benchmark/RaytraceBenchmark.cu