This uses a simple adaptation of the `LoopNavigator.h` (changed into just `Navigator.h` in the benchmark's directory) to use the BVH acceleration structure on the GPU. This requires VecGeom's latest tag (1.1.16) in order to work.

Here are reported run times (Release build in both cases):

**Before:** CPU:  168.456 s, GPU: 61.872 s
**After**: CPU: 4.405 s, GPU: 1.063 s

**Speedups:** CPU: ≈38.24x, GPU:  ≈58.2x

The resulting rendered images are identical as far as I can tell, so the BVH navigator is working on the GPU as expected.

For your convenience, here is a diff of the headers`LoopNavigator.h` and `Navigator.h`:
```diff
--- base/inc/AdePT/LoopNavigator.h	2021-05-21 11:46:24.473844224 +0200
+++ examples/Raytracer_Benchmark/Navigator.h	2021-06-02 14:22:50.814901356 +0200
@@ -2,17 +2,18 @@
 // SPDX-License-Identifier: Apache-2.0
 
 /**
- * @file LoopNavigator.h
+ * @file Navigator.h
  * @brief Navigation methods for geometry.
  */
 
-#ifndef RT_LOOP_NAVIGATOR_H_
-#define RT_LOOP_NAVIGATOR_H_
+#ifndef RT_NAVIGATOR_H_
+#define RT_NAVIGATOR_H_
 
 #include <CopCore/Global.h>
 
 #include <VecGeom/base/Global.h>
 #include <VecGeom/base/Vector3D.h>
+#include <VecGeom/management/BVHManager.h>
 #include <VecGeom/navigation/NavStateIndex.h>
 
 #ifdef VECGEOM_ENABLE_CUDA
@@ -21,7 +22,7 @@
 
 inline namespace COPCORE_IMPL {
 
-class LoopNavigator {
+class Navigator {
 
 public:
   using VPlacedVolumePtr_t = vecgeom::VPlacedVolume const *;
@@ -36,26 +37,22 @@
       if (!vol->UnplacedContains(point)) return nullptr;
     }
 
-    VPlacedVolumePtr_t currentvolume = vol;
+    path.Push(vol);
+
     vecgeom::Vector3D<vecgeom::Precision> currentpoint(point);
-    path.Push(currentvolume);
+    vecgeom::Vector3D<vecgeom::Precision> daughterlocalpoint;
 
-    bool godeeper;
-    do {
-      godeeper = false;
-      for (auto *daughter : currentvolume->GetDaughters()) {
-        vecgeom::Vector3D<vecgeom::Precision> transformedpoint;
-        if (daughter->Contains(currentpoint, transformedpoint)) {
-          path.Push(daughter);
-          currentpoint  = transformedpoint;
-          currentvolume = daughter;
-          godeeper      = true;
-          break;
-        }
-      }
-    } while (godeeper);
+    for (auto v = vol; v->GetDaughters().size() > 0;) {
+      auto bvh = vecgeom::BVHManager::GetBVH(v->GetLogicalVolume()->id());
+
+      if (!bvh->LevelLocate(currentpoint, v, daughterlocalpoint))
+        break;
 
-    return currentvolume;
+      currentpoint = daughterlocalpoint;
+      path.Push(v);
+    }
+
+    return path.Top();
   }
 
   __host__ __device__
@@ -97,15 +94,9 @@
 
     if (step < 0) step = 0;
 
-    for (auto *daughter : pvol->GetDaughters()) {
-      double ddistance = daughter->DistanceToIn(localpoint, localdir, step);
-
-      // if distance is negative; we are inside that daughter and should relocate
-      // unless distance is minus infinity
-      const bool valid = (ddistance < step && !vecgeom::IsInf(ddistance)) &&
-                         !((ddistance <= 0.) && in_state.GetLastExited() == daughter);
-      hitcandidate = valid ? daughter : hitcandidate;
-      step         = valid ? ddistance : step;
+    if (pvol->GetDaughters().size() > 0) {
+      auto bvh = vecgeom::BVHManager::GetBVH(pvol->GetLogicalVolume()->id());
+      bvh->CheckDaughterIntersections(localpoint, localdir, step, pvol, hitcandidate);
     }
 
     // now we have the candidates and we prepare the out_state
```