@amadio 2021-06-03
I will rename it and move to the same place as `LoopNavigator.h` then. I will see what I can do about the CI as well.