@hahnjo 2021-06-04
I get similar results with `trackML.gdml`, but `cms2018.gdml` crashes in `vecgeom::cuda::BVHManager::GetBVH`. I'm not sure though if the BVH is to blame because it also crashes on `master` without this PR, but somewhere else (in `vecgeom::cuda::NavStateIndex::NavIndAddr` because you cannot call that when compiling with `nvcc` for the host?).