@amadio 2021-06-04
Ok, I will try to reproduce the problem using cms2018. However, since the materials are hard-coded, how did you change the code to be able to use cms2018.gdml?