@amadio 2021-06-04
I think that the geometry may not be properly setup, given all the hard-coded stuff, so I didn't even think about trying with CMS. I will see what happens in any case.