@amadio 2021-06-04
I figured out what is happening, at least on the CPU. A call to the navigator is made before the initialization of the BVH structures...