@amadio 2021-06-04
On the GPU the problem is that there is a call to the navigator on the CPU, but using `vecgeom::cuda::` namespace so it's the wrong copy of the CPU code that cannot/should not be called. This is *again* due to the double compilation on the host and namespacing issues. The initialization of the BVH structures happens for the host version of the code in the `vecgeom::cxx` namespace, so if you call the code in the `vecgeom::cuda` namespace (which happens automatically from a `.cu` file, as the inline namespace there is the `vecgeom::cuda` namespace), it will try to access uninitialized variables, that's why `GetBVH` crashes. The problem is that with VecGeom headers flip flopping which namespace is inline depending on how you compile things, this is only going to get worse and worse, as the `vecgeom::cxx` namespace is not accessible from `.cu` files and `vecgeom::cuda` namespace is not accessible from `.cpp` files of external projects trying to use VecGeom. Funnily enough, I cannot use the `CudaCheckError` function from VecGeom on a `.cu` file in AdePT for that reason! So we keep having to wrap calls to VecGeom in functions (like the `InitBVH` function I added to the RaytraceBenchmark example) and placing their body on a `.cpp` file if you need to call things from `vecgeom::cxx` and on `.cu` file if you need to call `vecgeom::cuda` stuff, so that, for example the `vecgeom::cxx::BVHManager::DeviceInit()` function can be called somewhere in a `.cu` file to initialize the GPU BVH structures if needed. This makes development using VecGeom on GPUs **_very very_** painful and wastes lots of time of developers that have to debug this sort of thing.