@agheata 2021-06-04
> ... This makes development using VecGeom on GPUs **_very very_** painful and wastes lots of time of developers that have to debug this sort of thing.

You are right and we have seen how difficult it is to integrate a new structure/type in VecGeom and move it to the device. Also, this example is a demonstration of how complicated management of the namespaces becomes, in particular, if you want to have a single executable for host/device and common code.  I think we all agree, but we will have to still live with this schema for a while until we manage to refactor it