@amadio 2021-06-04
The question that remains is: how does it even work for trackML? I think that the answer lies here:
```

```
So, after I fixed the problems discussed above, I now get this on the GPU:
```
RaytraceBenchmark: /tmp/vecgeom/include/VecGeom/navigation/NavStateIndex.h:148: static const NavIndex_t* vecgeom::cuda::NavStateIndex::NavIndAddr(NavIndex_t): Assertion `false && "reached unimplement code"' failed.
```
I guess this is due to the navigation index table not being deep enough to cover all of the geometry.