@hahnjo 2021-06-04
> I guess this is due to the navigation index table not being deep enough to cover all of the geometry.

I *think* this may again be an instance of the host calling CUDA code, at least that's what the comment immediately above the `assert` suggests...