@amadio 2021-06-04
As a last resort, I also tried to just move the camera point outside the geometry and use the code as is (from this PR). This is what I get:
```
...
=== Allocating 93.438 MB of ray data on the BackendType::CUDA

CUDA Exception: Lane User Stack Overflow

Thread 1 "RaytraceBenchma" received signal CUDA_EXCEPTION_2, Lane User Stack Overflow.
[Switching focus to CUDA kernel 3, grid 60079, block (6,0,0), thread (123,0,0), device 0, sm 12, warp 3, lane 27]
0x0000555559bc3fd0 in vecgeom::cuda::Transformation3D::Transform<-1, -1, double> (this=<optimized out>, master=..., local=...)
    at /tmp/vecgeom/include/VecGeom/base/Transformation3D.h:661
661	void Transformation3D::Transform(Vector3D<InputType> const &master, Vector3D<InputType> &local) const
```
So, even if we fix the problems with calling code from the wrong namespace, it will probably not be enough to get something working with the full CMS geometry.