@agheata 2021-06-04
> So, even if we fix the problems with calling code from the wrong namespace, it will probably not be enough to get something working with the full CMS geometry.

Well, not necessary, but it's bad indeed that the code is stack-hungry. I've tried and succeeded in the past to increase the CUDA thread stack size, maybe we should try this first: https://gitlab.cern.ch/VecGeom/VecGeom/-/blob/35faafd3346978ea5d5bfd0fcfed1fd5c50828e8/source/benchmarking/RaytraceBenchmark.cu#L90-97