@amadio 2021-06-04
I think we may also need to add an initialization mode for the CMS geometry later on. This is what I get when the code "works" on the CPU (with default settings):
![output](https://user-images.githubusercontent.com/249404/120782900-c4717400-c52a-11eb-80f9-c4ddf0258667.png)